Embrace prop eq.
We want to optimize for ease of use in the abstract domain, not the
concrete domain. Also, some day HoTT is coming. In the mean time, if
we really need to quotient we can just pretend the non-standard forms
don't exist and we'll be fine. In the absolute worst case, we can
assume proof-irrelevance and pair our types with a proof of normal
form.